+++ migrations/20251021031457_change_user_id_to_uuid_in_wallets_table.ts
@@ -1,33 +1,41 @@
 import { Knex } from 'knex';
 
 export async function up(knex: Knex): Promise<void> {
-  return knex.schema.alterTable('wallets', (table) => {
-    table.dropForeign(['user_id']);
-    table.dropUnique(['user_id']);
-    table.uuid('user_id').alter();
-    table
-      .foreign('user_id')
-      .references('id')
-      .inTable('users')
-      .onDelete('CASCADE');
-    table.unique(['user_id']);
-  });
+  const hasUserId = await knex.schema.hasColumn('wallets', 'user_id');
+  if (hasUserId) {
+    await knex.schema.alterTable('wallets', (table) => {
+      table.dropForeign('user_id', 'wallets_user_id_foreign');
+      table.dropUnique(['user_id']);
+    });
+
+    await knex.schema.alterTable('wallets', (table) => {
+      table.uuid('user_id').alter();
+    });
+
+    await knex.schema.alterTable('wallets', (table) => {
+      table
+        .foreign('user_id', 'wallets_user_id_foreign')
+        .references('id')
+        .inTable('users')
+        .onDelete('CASCADE');
+      table.unique(['user_id']);
+    });
+  }
 }
 
-/**
- * @param { import("knex").Knex } knex
- * @returns { Promise<void> }
- */
 export async function down(knex: Knex): Promise<void> {
-  return knex.schema.alterTable('wallets', (table) => {
-    table.dropForeign(['user_id']);
-    table.dropUnique(['user_id']);
-    table.integer('user_id').alter();
-    table
-      .foreign('user_id')
-      .references('id')
-      .inTable('users')
-      .onDelete('CASCADE');
-    table.unique(['user_id']);
-  });
+  const hasUserId = await knex.schema.hasColumn('wallets', 'user_id');
+  if (hasUserId) {
+    await knex.schema.alterTable('wallets', (table) => {
+      table.dropForeign('user_id', 'wallets_user_id_foreign');
+      table.dropUnique(['user_id']);
+      table.integer('user_id').alter();
+      table
+        .foreign('user_id', 'wallets_user_id_foreign')
+        .references('id')
+        .inTable('users')
+        .onDelete('CASCADE');
+      table.unique(['user_id']);
+    });
+  }
 }
